enic: added enic_wq.c and enic_wq.h
jira LE-4311
Rebuild_History Non-Buildable kernel-5.14.0-570.46.1.el9_6
commit-author Satish Kharat <satishkh@cisco.com>
commit e5f1bcd93d96b338c486f2376f260f7fc98b3cd5

Moves wq related function to enic_wq.c. Prepares for
a cleaup of enic wq code path.

Co-developed-by: Nelson Escobar <neescoba@cisco.com>
Co-developed-by: John Daley <johndale@cisco.com>
Link: https://patch.msgid.link/20250304-enic_cleanup_and_ext_cq-v2-6-85804263dad8@cisco.com
(cherry picked from commit e5f1bcd93d96b338c486f2376f260f7fc98b3cd5)
